requirements/base.txt
Django==1.9.5
psycopg2

celery
django-celery

# PEP-8 conventions
flake8
pep8-naming

# Utilities
ipython
